Add gatewaymac to the encrypted query string
gatewaymac can be useful for identifying the gateway on the remote FAS.

gatewayname is intended to be used for this purpose,
but in the case of a rollout of a standard build, gatewaymac could
also be useful.

The variable $gatewaymac has been added to the remote FAS example files,
fas-aes.php and fas-aes-https.php.

This variable is only passed for fas_secure_enabled levels 2 and 3

Note: The value gatewaymac is the mac address of the gateway interface
and is not guaranteed to be the mac address displayed on the label.

Signed-off-by: Rob White <rob@blue-wave.net>