Add: Support for Custom Parameters This enhancement allows custom parameters to be defined in the config file.
This enhancement is added primarily to support remote configuration operations
using tools such as opensync, but can also be of general use.

Custom Parameters are options defined in the configuration file
and have fixed values once set.

Note: Custom Variables are very different and are defined in the FAS,
with values determined by the FAS/Client dialogue script.

Signed-off-by: Rob White <rob@blue-wave.net>