FAS: Add base64 encoding to fas_secure_enabled level 1
This makes passing parameters and variables much easier in a client/fas
login dialogue as well as passing them to the landing and status pages.

It also provides a higher degree of security as ip and mac addresses
are no longer sent in clear text although could be decoded easy enough.

Signed-off-by: Rob White <rob@blue-wave.net>