ndsctl: Fix ndsctl locking by using lockf()
There are currently a few issues with the current locking approach:

For one thing, if ndsctl is terminated while it is running it will leave
a stale lock file which will lead to a failure of any subsequent ndsctl
invocation. Which can also lead to failing splash page access if the
example PreAuth script is used.

For another the moment in ndsctl or openNDS between checking if a lock
file exists and actually creating it can race with concurrent instances
trying to get/create this lock.

Furthermore when the lock is held by another instance instead of waiting
a brief moment for the lock to become free an error is displayed
instead. Which can lead to temporary failures of binauth, for example.

Instead an OS assisted locking mechanism should be used. Like lockf().
It has the following advantages:

When a process dies the lock is freed inderectly by the OS, as the
OS will automatically close the open file handle. And it is race free
regarding checking and obtaining the lock.

The deletion of the lock file is not necessary anymore and in fact
discouraged when using lockf(). As lockf() does not lock/unlock by
creating/removing the file but instead by using a POSIX lock on (a
section of) the file.

While switching to lockf() also adding retries. For up to one second
ndsctl or openNDS will check the availability of the lock file every
millisecond. And will exit with an error only after this one second
timeout.

This change (next to others) is also necessary to be able to run
multiple openNDS instances simultaneously later.

Signed-off-by: Linus Lüssing <ll@simonwunderlich.de>