ndsctl: remove superfluous space transmission in "ndsctl json"
When a command is send over netlink with no parameter it should be just
send like:

  "json\r\n\r\n"

Howver the previous "if (param)" check is bogus, as param will always
point to the address on the stack of the args array in main() and will
therefore never be a NULL pointer.

Instead we should check if param (= args) is an empty string. We can
simply do that by dereferencing the first byte of the array and checking
if it is 0 (= a '\0' character).

Fixes: d47f1fe ("ndsctl: refactor code to reduce complexity")
Signed-off-by: Linus Lüssing <ll@simonwunderlich.de>