Fix - Preemptive auth failure after previous deauth.
When a client was deauthed, it was not removed from the client list,
instead its firewall state was changed from authenticated to preauthenticated.
With the introduction of preemptive authentication, this is no longer sufficient.

If a client had been deauthed for any reason, then was pre-emptively re-authed
(within the configured preauthidletimeout period),
it would inherit its previous details including its previous ip address
as far as openNDS is concerned.
If the dhcp allocated ip address was the same as before then there was no problem.

But if that previous ip address had been leased to some other client then
the preemptive authentication would fail and the other client would be locked out.

This fix deletes the client from the client list and does more in depth
checking for the presence of the client before preemtive auth is allowed.

Signed-off-by: Rob White <rob@blue-wave.net>